requirements.txt
opencv-python==4.12.0.88
carla==0.9.16
numpy==1.26.4
pyquaternion==0.9.9
pyyaml==5.4.1
tqdm==4.67.1
